Move top search form to the separate component
