+++ src/components/layout-header-search.jsx
@@ -1,0 +1,173 @@
+import cn from 'classnames';
+import { faSearch, faSlidersH, faTimesCircle } from '@fortawesome/free-solid-svg-icons';
+import { useState, useEffect, useRef, useMemo } from 'react';
+import { useEvent } from 'react-use-event-hook';
+import { withRouter } from 'react-router';
+import { KEY_ESCAPE } from 'keycode-js';
+import styles from './layout-header.module.scss';
+import { Icon } from './fontawesome-icons';
+import { Autocomplete } from './autocomplete/autocomplete';
+import { ButtonLink } from './button-link';
+import { useMediaQuery } from './hooks/media-query';
+
+const autocompleteAnchor = /(^|[^a-z\d])@|((from|to|author|by|in|commented-?by|liked-?by):)/gi;
+
+export const HeaderSearchForm = withRouter(function HeaderSearchForm({ router, closeSearchForm }) {
+  const isWideScreen = useMediaQuery('(min-width: 700px)');
+  const isNarrowScreen = useMediaQuery('(max-width: 549px)');
+  const onSearchPage = router.routes[router.routes.length - 1].name === 'search';
+
+  const [searchExpanded, setSearchExpanded] = useState(false);
+  const [query, setQuery] = useState('');
+  const onQueryChange = useEvent(({ target }) => setQuery(target.value));
+
+  const fullSearchForm = isWideScreen;
+  const compactSearchForm = !fullSearchForm;
+  const collapsibleSearchForm = isNarrowScreen && (!onSearchPage || searchExpanded);
+
+  useEffect(() => {
+    if (!collapsibleSearchForm) {
+      setSearchExpanded(false);
+    }
+  }, [collapsibleSearchForm]);
+
+  const initialQuery = useInitialQuery(router);
+  const input = useRef(null);
+  useEffect(() => void setQuery(initialQuery), [initialQuery]);
+
+  const onSubmit = useEvent((e) => {
+    e.preventDefault();
+    const q = query.trim();
+    if (q !== '') {
+      router.push(`/search?q=${encodeURIComponent(q)}`);
+      input.current.blur();
+    }
+  });
+
+  const showAdvancedSearch = useEvent(() => {
+    router.push(`/search?q=${encodeURIComponent(query.trim())}&advanced`);
+    input.current.blur();
+  });
+
+  const onFocus = useEvent(() => isNarrowScreen && onSearchPage && setSearchExpanded(true));
+  const onKeyDown = useEvent((e) => e.keyCode === KEY_ESCAPE && input.current.blur());
+  const clearSearchForm = useEvent(() => (setQuery(''), input.current.focus()));
+
+  const focusHandlers = useDebouncedFocus({
+    onFocus,
+    onBlur: closeSearchForm,
+  });
+
+  useEffect(() => {
+    const abortController = new AbortController();
+    document.addEventListener(
+      'keydown',
+      (e) => {
+        if (document.activeElement !== document.body) {
+          return;
+        }
+        if (
+          // [/] or Ctrl+[K]
+          (e.code === 'Slash' && !e.ctrlKey && !e.altKey && !e.metaKey && !e.shiftKey) ||
+          (e.code === 'KeyK' && e.ctrlKey && !e.altKey && !e.metaKey && !e.shiftKey)
+        ) {
+          input.current.focus();
+          e.preventDefault();
+        }
+      },
+      { signal: abortController.signal },
+    );
+    return () => abortController.abort();
+  }, []);
+
+  return (
+    <form className={styles.searchForm} action="/search" onSubmit={onSubmit}>
+      <span className={styles.searchInputContainer} {...focusHandlers} tabIndex={0}>
+        <span className={styles.searchInputBox}>
+          <input
+            className={styles.searchInput}
+            type="text"
+            name="q"
+            ref={input}
+            placeholder={isWideScreen ? 'Press / to search' : ''}
+            autoFocus={collapsibleSearchForm}
+            autoComplete="off"
+            value={query}
+            onChange={onQueryChange}
+            onKeyDown={onKeyDown}
+            tabIndex={-1}
+          />
+          <button
+            type="button"
+            className={cn(styles.clearSearchButton, styles.compactButton)}
+            aria-label="Clear search form"
+            title="Clear search form"
+            onClick={clearSearchForm}
+            tabIndex={-1}
+          >
+            <Icon icon={faTimesCircle} />
+          </button>
+          <div className={styles.autocompleteBox}>
+            <ButtonLink tag="div" className={styles.advancedSearch} onClick={showAdvancedSearch}>
+              <Icon icon={faSlidersH} className={styles.advancedSearchIcon} />
+              <span>Advanced search options</span>
+            </ButtonLink>
+            <Autocomplete inputRef={input} context="search" anchor={autocompleteAnchor} />
+          </div>
+        </span>
+        {compactSearchForm && <Icon icon={faSearch} className={styles.searchIcon} />}
+      </span>
+      {fullSearchForm && (
+        <button type="submit" className={styles.searchButton}>
+          Search
+        </button>
+      )}
+    </form>
+  );
+});
+
+function useInitialQuery(router) {
+  return useMemo(() => {
+    const route = router.routes[router.routes.length - 1];
+    switch (route.name) {
+      case 'search':
+        return (router.location.query.q || router.location.query.qs || '').trim();
+      case 'saves':
+        return `in-my:saves `;
+      case 'discussions':
+        return `in-my:discussions `;
+      case 'direct':
+        return `in-my:directs `;
+      case 'userLikes':
+        return `liked-by:${router.params.userName} `;
+      case 'userComments':
+        return `commented-by:${router.params.userName} `;
+      case 'userFeed':
+        return `in:${router.params.userName} `;
+      default:
+        return '';
+    }
+  }, [router.routes, router.params, router.location]);
+}
+
+function useDebouncedFocus({ onFocus: onFocusOrig, onBlur: onBlurOrig }, interval = 100) {
+  const focusTimer = useRef(0);
+  const blurTimer = useRef(0);
+
+  const cleanup = useEvent(() => {
+    window.clearTimeout(blurTimer.current);
+    window.clearTimeout(focusTimer.current);
+  });
+  useEffect(() => () => cleanup(), [cleanup]);
+
+  const onFocus = useEvent(() => {
+    cleanup();
+    focusTimer.current = window.setTimeout(onFocusOrig, interval);
+  });
+  const onBlur = useEvent(() => {
+    cleanup();
+    blurTimer.current = window.setTimeout(onBlurOrig, interval);
+  });
+
+  return { onFocus, onBlur };
+}
